+++ src/consul/src/ConfigProvider.php
@@ -17,12 +17,6 @@ class ConfigProvider
     public function __invoke(): array
     {
         return [
-            'dependencies' => [
-            ],
-            'scan' => [
-                'paths' => [
-                ],
-            ],
             'publish' => [
                 [
                     'id' => 'config',