+++ src/model-listener/src/ConfigProvider.php
@@ -19,16 +19,14 @@ class ConfigProvider
     public function __invoke(): array
     {
         return [
-            'dependencies' => [
-            ],
-            'commands' => [
-            ],
-            'scan' => [
-                'paths' => [
-                    __DIR__,
-                ],
-                'collectors' => [
-                    ListenerCollector::class,
+            'annotations' => [
+                'scan' => [
+                    'paths' => [
+                        __DIR__,
+                    ],
+                    'collectors' => [
+                        ListenerCollector::class,
+                    ],
                 ],
             ],
         ];