+++ CHANGELOG.md
@@ -88,6 +88,7 @@ Now:
 ## Fixed
 
 - [#448](https://github.com/hyperf-cloud/hyperf/pull/448) Fixed TCP Server does not works when HTTP Server or WebSocket Server exists.
+- [#647](https://github.com/hyperf-cloud/hyperf/pull/647) Append `eof` to TCP response, according to the server configuration.
 
 # v1.0.16 - 2019-09-20